Cast problems (?)
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@29574 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR0/GMMR0.cpp
@@ -2649,7 +2649,7 @@ DECLINLINE(void) gmmR0ConvertToSharedPage(PGMM pGMM, PGVM pGVM, RTHCPHYS HCPhys,
     pGVM->gmm.s.cPrivatePages--;
 
     /* Modify the page structure. */
-    pPage->Shared.pfn     = (uint32_t)(HCPhys >> PAGE_SHIFT);
+    pPage->Shared.pfn     = (uint32_t)(uint64_t)(HCPhys >> PAGE_SHIFT);
     pPage->Shared.cRefs   = 1;
     pPage->Common.u2State = GMM_PAGE_STATE_SHARED;
 }
@@ -3820,7 +3820,7 @@ new_shared_page:
                     goto new_shared_page; /* ugly goto */
                 }
 
-                Log(("Replace existing page guest %RGp host %RHp -> %RHp\n", paPageDesc[i].GCPhys, paPageDesc[i].HCPhys, pPage->Shared.pfn << PAGE_SHIFT));
+                Log(("Replace existing page guest %RGp host %RHp -> %RHp\n", paPageDesc[i].GCPhys, paPageDesc[i].HCPhys, ((uint64_t)pPage->Shared.pfn) << PAGE_SHIFT));
 
                 /* Calculate the virtual address of the local page. */
                 pChunk = gmmR0GetChunk(pGMM, paPageDesc[i].uHCPhysPageId >> GMM_CHUNKID_SHIFT);
@@ -3877,7 +3877,7 @@ new_shared_page:
                 gmmR0UseSharedPage(pGMM, pGVM, pPage);
 
                 /* Pass along the new physical address & page id. */
-                paPageDesc[i].HCPhys        = pPage->Shared.pfn << PAGE_SHIFT;
+                paPageDesc[i].HCPhys        = ((uint64_t)pPage->Shared.pfn) << PAGE_SHIFT;
                 paPageDesc[i].uHCPhysPageId = pGlobalRegion->paHCPhysPageID[i];
             }
         }
